pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "Evaluacion-modulo-10---BBDD-Vectoriales"
version = "0.1.0"
description = "End2EndFlow"
readme = "README.md"
requires-python = ">=3.11,<3.13"
dependencies = [
    "fastapi>=0.117.1,<1.0.0",
    "google-genai>=1.38.0",
    "llama-index>=0.14.2",
    "llama-index-embeddings-google-genai>=0.3.1",
    "llama-index-llms-google-genai>=0.5.1",
    "llama-index-vector-stores-qdrant>=0.8.5",
    "python-dotenv>=1.0.1",
    "pytest==8.2.0",
    "pypdf>=6.1.0",
    "qdrant-client==1.15.1",
    "ruff==0.4.3",
    "uvicorn[standard]>=0.30.0,<1.0.0",
]

[tool.pytest.ini_options]
pythonpath = ["."]
